refactor main page components for modular structure
