+++ frontend/src/app/page.tsx
@@ -1,631 +1,75 @@
 'use client';
 
-import { useEffect, useState, useRef } from 'react';
-
-interface Message {
-  id: string;
-  role: 'user' | 'assistant';
-  content: string;
-  timestamp: Date;
-}
+import { Header } from '@/components/Header';
+import { ResumeUpload } from '@/components/ResumeUpload';
+import { HowItWorks } from '@/components/HowItWorks';
+import { ChatContainer } from '@/components/Chat/ChatContainer';
+import { useTheme } from '@/hooks/useTheme';
+import { useFileUpload } from '@/hooks/useFileUpload';
+import { useChat } from '@/hooks/useChat';
 
 export default function Home() {
-  const [isDark, setIsDark] = useState(false);
-  const [uploadedFile, setUploadedFile] = useState<File | null>(null);
-  const [isDragging, setIsDragging] = useState(false);
-  const [uploadError, setUploadError] = useState<string | null>(null);
-  const [messages, setMessages] = useState<Message[]>([]);
-  const [inputText, setInputText] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const textareaRef = useRef<HTMLTextAreaElement>(null);
-  const messagesEndRef = useRef<HTMLDivElement>(null);
-
-  // Initialize theme from localStorage
-  useEffect(() => {
-    const savedTheme = localStorage.getItem('theme');
-    const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
-    const shouldBeDark = savedTheme === 'dark' || (!savedTheme && prefersDark);
-    setIsDark(shouldBeDark);
-    
-    if (shouldBeDark) {
-      document.documentElement.classList.add('dark');
-    } else {
-      document.documentElement.classList.remove('dark');
-    }
-  }, []);
-
-  const toggleTheme = () => {
-    const newTheme = !isDark;
-    setIsDark(newTheme);
-    localStorage.setItem('theme', newTheme ? 'dark' : 'light');
-    
-    if (newTheme) {
-      document.documentElement.classList.add('dark');
-    } else {
-      document.documentElement.classList.remove('dark');
-    }
-  };
-
-  const validateFile = (file: File): string | null => {
-    // Check file type
-    if (file.type !== 'application/pdf') {
-      return 'Please upload a PDF file only.';
-    }
-    
-    // Check file size (10MB = 10 * 1024 * 1024 bytes)
-    const maxSize = 10 * 1024 * 1024;
-    if (file.size > maxSize) {
-      return 'File size must be less than 10MB.';
-    }
-    
-    return null;
-  };
-
-  const handleFileSelect = (file: File) => {
-    setUploadError(null);
-    const error = validateFile(file);
-    
-    if (error) {
-      setUploadError(error);
-      return;
-    }
-    
-    setUploadedFile(file);
-  };
-
-  const handleDragOver = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  };
-
-  const handleDragLeave = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-  };
-
-  const handleDrop = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-
-    const files = e.dataTransfer.files;
-    if (files && files.length > 0) {
-      handleFileSelect(files[0]);
-    }
-  };
-
-  const handleFileInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const files = e.target.files;
-    if (files && files.length > 0) {
-      handleFileSelect(files[0]);
-    }
-  };
-
-  const handleClickUpload = () => {
-    fileInputRef.current?.click();
-  };
-
-  const handleRemoveFile = () => {
-    setUploadedFile(null);
-    setUploadError(null);
-    if (fileInputRef.current) {
-      fileInputRef.current.value = '';
-    }
-  };
-
-  const formatFileSize = (bytes: number): string => {
-    if (bytes < 1024) return bytes + ' B';
-    if (bytes < 1024 * 1024) return (bytes / 1024).toFixed(2) + ' KB';
-    return (bytes / (1024 * 1024)).toFixed(2) + ' MB';
-  };
-
-  // Auto-scroll to bottom when new messages arrive
-  useEffect(() => {
-    messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
-  }, [messages]);
-
-  // Auto-resize textarea
-  useEffect(() => {
-    if (textareaRef.current) {
-      textareaRef.current.style.height = 'auto';
-      textareaRef.current.style.height = `${Math.min(textareaRef.current.scrollHeight, 200)}px`;
-    }
-  }, [inputText]);
-
-  const getAgentIntroduction = (): string => {
-    return `Hello! I'm your Resume AI Agent. I'm here to help you optimize your resume to perfectly match job descriptions.
-
-Here's what I can do for you:
-• Analyze your resume and identify areas for improvement
-• Match your skills and experience to specific job requirements
-• Suggest targeted improvements to make your resume stand out
-• Help you tailor your resume for different positions
-
-To get started, please upload your resume PDF and share a job description (either by pasting the text or providing a URL). I'll then analyze both and provide personalized recommendations to help you land that interview!
-
-What would you like to work on today?`;
-  };
-
-  const handleSendMessage = async () => {
-    if (!inputText.trim() || isLoading) return;
-
-    const userMessage: Message = {
-      id: Date.now().toString(),
-      role: 'user',
-      content: inputText.trim(),
-      timestamp: new Date(),
-    };
-
-    setMessages((prev) => [...prev, userMessage]);
-    setInputText('');
-    setIsLoading(true);
-
-    // Simulate API call delay
-    setTimeout(() => {
-      const assistantMessage: Message = {
-        id: (Date.now() + 1).toString(),
-        role: 'assistant',
-        content: getAgentIntroduction(),
-        timestamp: new Date(),
-      };
-
-      setMessages((prev) => [...prev, assistantMessage]);
-      setIsLoading(false);
-    }, 1000);
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
-    if (e.key === 'Enter' && !e.shiftKey) {
-      e.preventDefault();
-      handleSendMessage();
-    }
-  };
+  const { isDark, toggleTheme } = useTheme();
+  const {
+    uploadedFile,
+    isDragging,
+    uploadError,
+    fileInputRef,
+    handleDragOver,
+    handleDragLeave,
+    handleDrop,
+    handleFileInputChange,
+    handleClickUpload,
+    handleRemoveFile,
+  } = useFileUpload();
+  const {
+    messages,
+    inputText,
+    setInputText,
+    isLoading,
+    textareaRef,
+    messagesEndRef,
+    handleSendMessage,
+    handleKeyDown,
+  } = useChat();
 
   return (
     <div className="flex min-h-screen flex-col bg-gradient-to-br from-slate-50 via-blue-50 to-indigo-50 dark:from-slate-950 dark:via-slate-900 dark:to-slate-950">
-      {/* Header */}
-      <header className="sticky top-0 z-50 w-full border-b border-slate-200/80 bg-white/80 backdrop-blur-md dark:border-slate-700/80 dark:bg-slate-900/80">
-        <div className="relative flex h-16 w-full items-center">
-          <div className="container mx-auto flex h-16 items-center justify-between px-6">
-            <div className="flex items-center gap-3">
-              <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-gradient-to-br from-indigo-600 to-blue-600 dark:from-indigo-500 dark:to-blue-500">
-                <svg className="h-6 w-6 text-white" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" />
-                </svg>
-              </div>
-              <h1 className="text-xl font-semibold tracking-tight text-slate-900 dark:text-slate-100">
-                Resume AI Agent
-              </h1>
-            </div>
-          </div>
-          
-          {/* Dark Mode Toggle - Positioned at absolute top-right */}
-          <div className="absolute right-6 flex items-center gap-4">
-            <button
-              onClick={toggleTheme}
-              className="group relative h-9 w-9 rounded-lg border border-slate-200 bg-slate-50 transition-all hover:bg-slate-100 dark:border-slate-700 dark:bg-slate-800 dark:hover:bg-slate-700"
-              aria-label="Toggle theme"
-            >
-              {/* Sun Icon (visible in dark mode) */}
-              <svg
-                className={`absolute left-1/2 top-1/2 h-5 w-5 -translate-x-1/2 -translate-y-1/2 text-slate-600 transition-all duration-300 dark:text-amber-400 ${
-                  isDark ? 'rotate-0 scale-100 opacity-100' : 'rotate-90 scale-0 opacity-0'
-                }`}
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
-                />
-              </svg>
-              
-              {/* Moon Icon (visible in light mode) */}
-              <svg
-                className={`absolute left-1/2 top-1/2 h-5 w-5 -translate-x-1/2 -translate-y-1/2 text-slate-600 transition-all duration-300 dark:text-slate-400 ${
-                  isDark ? '-rotate-90 scale-0 opacity-0' : 'rotate-0 scale-100 opacity-100'
-                }`}
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z"
-                />
-              </svg>
-            </button>
-          </div>
-        </div>
-      </header>
+      <Header isDark={isDark} toggleTheme={toggleTheme} />
 
       {/* Main Content Area */}
       <main className="container mx-auto flex flex-1 flex-col gap-6 p-6">
         <div className="flex flex-1 flex-col gap-6 lg:flex-row">
           {/* Left Sidebar - Upload Section */}
           <aside className="flex flex-col gap-4 lg:w-80">
-            <div className="rounded-xl border border-slate-200 bg-white p-6 shadow-sm dark:border-slate-700 dark:bg-slate-800">
-              <h2 className="mb-4 text-lg font-semibold text-slate-900 dark:text-slate-100">
-                Upload Resume
-              </h2>
-              
-              {/* Hidden file input */}
-              <input
-                ref={fileInputRef}
-                type="file"
-                accept=".pdf,application/pdf"
-                onChange={handleFileInputChange}
-                className="hidden"
-              />
-              
-              {/* PDF Upload Area */}
-              {!uploadedFile ? (
-                <div
+            <ResumeUpload
+              uploadedFile={uploadedFile}
+              isDragging={isDragging}
+              uploadError={uploadError}
+              fileInputRef={fileInputRef}
                   onDragOver={handleDragOver}
                   onDragLeave={handleDragLeave}
                   onDrop={handleDrop}
-                  onClick={handleClickUpload}
-                  className={`flex h-40 cursor-pointer flex-col items-center justify-center rounded-lg border-2 border-dashed p-6 text-center transition-all ${
-                    isDragging
-                      ? 'border-indigo-500 bg-indigo-100 dark:border-indigo-400 dark:bg-indigo-950/50'
-                      : 'border-slate-300 bg-slate-50 hover:border-indigo-400 hover:bg-indigo-50/50 dark:border-slate-600 dark:bg-slate-700/50 dark:hover:border-indigo-500 dark:hover:bg-indigo-950/30'
-                  }`}
-                >
-                  <svg
-                    className={`mb-3 h-10 w-10 transition-colors ${
-                      isDragging
-                        ? 'text-indigo-600 dark:text-indigo-400'
-                        : 'text-slate-400 dark:text-slate-500'
-                    }`}
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12"
-                    />
-                  </svg>
-                  <p className="text-sm font-medium text-slate-600 dark:text-slate-300">
-                    {isDragging ? 'Drop PDF here' : 'Drop PDF here or click to browse'}
-                  </p>
-                  <p className="mt-1 text-xs text-slate-500 dark:text-slate-400">
-                    Max file size: 10MB
-                  </p>
-                </div>
-              ) : (
-                <div className="rounded-lg border border-slate-200 bg-slate-50 p-4 dark:border-slate-700 dark:bg-slate-700/50">
-                  <div className="flex items-start gap-3">
-                    <div className="flex h-10 w-10 flex-shrink-0 items-center justify-center rounded-lg bg-indigo-100 dark:bg-indigo-900/50">
-                      <svg
-                        className="h-6 w-6 text-indigo-600 dark:text-indigo-400"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        stroke="currentColor"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M7 21h10a2 2 0 002-2V9.414a1 1 0 00-.293-.707l-5.414-5.414A1 1 0 0012.586 3H7a2 2 0 00-2 2v14a2 2 0 002 2z"
+              onFileInputChange={handleFileInputChange}
+              onClickUpload={handleClickUpload}
+              onRemoveFile={handleRemoveFile}
                         />
-                      </svg>
-                    </div>
-                    <div className="min-w-0 flex-1">
-                      <p className="truncate text-sm font-medium text-slate-900 dark:text-slate-100">
-                        {uploadedFile.name}
-                      </p>
-                      <p className="mt-1 text-xs text-slate-500 dark:text-slate-400">
-                        {formatFileSize(uploadedFile.size)}
-                      </p>
-                    </div>
-                    <button
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        handleRemoveFile();
-                      }}
-                      className="flex-shrink-0 rounded-lg p-1 text-slate-400 transition-colors hover:bg-slate-200 hover:text-slate-600 dark:hover:bg-slate-600 dark:hover:text-slate-300"
-                      aria-label="Remove file"
-                    >
-                      <svg
-                        className="h-5 w-5"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        stroke="currentColor"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M6 18L18 6M6 6l12 12"
-                        />
-                      </svg>
-                    </button>
-                  </div>
-                </div>
-              )}
-              
-              {/* Error Message */}
-              {uploadError && (
-                <div className="mt-3 rounded-lg border border-red-200 bg-red-50 p-3 dark:border-red-900/50 dark:bg-red-950/30">
-                  <div className="flex items-start gap-2">
-                    <svg
-                      className="h-5 w-5 flex-shrink-0 text-red-600 dark:text-red-400"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-                      />
-                    </svg>
-                    <p className="text-sm text-red-700 dark:text-red-400">{uploadError}</p>
-                  </div>
-                </div>
-              )}
-            </div>
-
-            {/* Instructions Card */}
-            <div className="rounded-xl border border-indigo-200 bg-gradient-to-br from-indigo-50 to-blue-50 p-5 dark:border-indigo-900/50 dark:from-indigo-950/50 dark:to-blue-950/50">
-              <h3 className="mb-2 text-sm font-semibold text-indigo-900 dark:text-indigo-300">
-                How it works
-              </h3>
-              <ol className="space-y-2 text-sm text-slate-700 dark:text-slate-300">
-                <li className="flex gap-2">
-                  <span className="font-semibold text-indigo-600 dark:text-indigo-400">1.</span>
-                  <span>Upload your resume PDF</span>
-                </li>
-                <li className="flex gap-2">
-                  <span className="font-semibold text-indigo-600 dark:text-indigo-400">2.</span>
-                  <span>Paste job description or URL</span>
-                </li>
-                <li className="flex gap-2">
-                  <span className="font-semibold text-indigo-600 dark:text-indigo-400">3.</span>
-                  <span>Review AI suggestions</span>
-                </li>
-                <li className="flex gap-2">
-                  <span className="font-semibold text-indigo-600 dark:text-indigo-400">4.</span>
-                  <span>Download optimized resume</span>
-                </li>
-              </ol>
-            </div>
+            <HowItWorks />
           </aside>
 
           {/* Main Chat Area */}
           <div className="flex flex-1 flex-col gap-4">
-            {messages.length === 0 ? (
-              /* Empty State - Improved Layout for Large Screens */
-              <div className="flex flex-1 flex-col rounded-xl border border-slate-200 bg-white shadow-sm dark:border-slate-700 dark:bg-slate-800">
-                <div className="flex flex-1 flex-col items-center justify-start pt-16 sm:pt-20 md:pt-24 lg:pt-32">
-                  <div className="mx-auto w-full max-w-3xl px-6">
-                    {/* Question */}
-                    <div className="mb-10 text-center">
-                      <h2 className="mb-3 text-3xl font-semibold text-slate-900 dark:text-slate-100 sm:text-4xl">
-                        What would you like to work on today?
-                      </h2>
-                      <p className="text-sm text-slate-500 dark:text-slate-400 sm:text-base">
-                        Upload your resume and ask me anything to get started
-                      </p>
-                    </div>
-                    
-                    {/* Large Input Field */}
-                    <div className="w-full">
-                      <div className="relative flex items-center rounded-2xl border border-slate-300 bg-slate-50 px-4 py-4 shadow-sm transition-all focus-within:border-indigo-500 focus-within:bg-white focus-within:shadow-md dark:border-slate-600 dark:bg-slate-700/50 dark:focus-within:border-indigo-500 dark:focus-within:bg-slate-700">
-                        {/* Plus Icon */}
-                        <button
-                          onClick={handleClickUpload}
-                          className="mr-3 flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-lg text-slate-400 transition-colors hover:bg-slate-200 hover:text-slate-600 dark:text-slate-500 dark:hover:bg-slate-600 dark:hover:text-slate-300"
-                          aria-label="Add attachment"
-                        >
-                          <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 4v16m8-8H4" />
-                          </svg>
-                        </button>
-                        
-                        {/* Text Input */}
-                        <textarea
-                          ref={textareaRef}
-                          value={inputText}
-                          onChange={(e) => setInputText(e.target.value)}
+            <ChatContainer
+              messages={messages}
+              inputText={inputText}
+              isLoading={isLoading}
+              textareaRef={textareaRef}
+              messagesEndRef={messagesEndRef}
+              onInputChange={setInputText}
                           onKeyDown={handleKeyDown}
-                          placeholder="Ask any question"
-                          className="flex-1 resize-none border-0 bg-transparent text-base text-slate-900 placeholder-slate-400 focus:outline-none focus:ring-0 dark:text-slate-100 dark:placeholder-slate-500"
-                          rows={1}
-                          style={{ minHeight: '32px', maxHeight: '200px' }}
-                        />
-                        
-                        {/* Right Side Icons */}
-                        <div className="ml-3 flex items-center gap-2">
-                          {/* Microphone Icon */}
-                          <button
-                            className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-lg text-slate-400 transition-colors hover:bg-slate-200 hover:text-slate-600 dark:text-slate-500 dark:hover:bg-slate-600 dark:hover:text-slate-300"
-                            aria-label="Voice input"
-                          >
-                            <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 11a7 7 0 01-7 7m0 0a7 7 0 01-7-7m7 7v4m0 0H8m4 0h4m-4-8a3 3 0 01-3-3V5a3 3 0 116 0v6a3 3 0 01-3 3z" />
-                            </svg>
-                          </button>
-                          
-                          {/* Audio Waveform Icon (Send Button) */}
-                          <button
-                            onClick={handleSendMessage}
-                            disabled={!inputText.trim() || isLoading}
-                            className={`flex h-9 w-9 flex-shrink-0 items-center justify-center rounded-full bg-indigo-600 text-white transition-all hover:bg-indigo-700 disabled:cursor-not-allowed disabled:opacity-50 dark:bg-indigo-500 dark:hover:bg-indigo-600 ${
-                              inputText.trim() && !isLoading ? 'hover:scale-105' : ''
-                            }`}
-                            aria-label="Send message"
-                          >
-                            {isLoading ? (
-                              <svg className="h-4 w-4 animate-spin" fill="none" viewBox="0 0 24 24">
-                                <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                                <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                              </svg>
-                            ) : (
-                              <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 19l9 2-9-18-9 18 9-2zm0 0v-8" />
-                              </svg>
-                            )}
-                          </button>
-                        </div>
-                      </div>
-                    </div>
-                    
-                    {/* Quick Suggestions - Only show on larger screens */}
-                    <div className="mt-8 hidden grid-cols-1 gap-3 sm:grid sm:grid-cols-2 lg:grid-cols-3">
-                      <button
-                        onClick={() => setInputText("How can I improve my resume?")}
-                        className="group rounded-lg border border-slate-200 bg-slate-50 px-4 py-3 text-left text-sm text-slate-700 transition-all hover:border-indigo-300 hover:bg-indigo-50 hover:text-indigo-700 dark:border-slate-700 dark:bg-slate-800/50 dark:text-slate-300 dark:hover:border-indigo-600 dark:hover:bg-indigo-950/30 dark:hover:text-indigo-400"
-                      >
-                        <span className="font-medium">💡</span> How can I improve my resume?
-                      </button>
-                      <button
-                        onClick={() => setInputText("Match my resume to a job description")}
-                        className="group rounded-lg border border-slate-200 bg-slate-50 px-4 py-3 text-left text-sm text-slate-700 transition-all hover:border-indigo-300 hover:bg-indigo-50 hover:text-indigo-700 dark:border-slate-700 dark:bg-slate-800/50 dark:text-slate-300 dark:hover:border-indigo-600 dark:hover:bg-indigo-950/30 dark:hover:text-indigo-400"
-                      >
-                        <span className="font-medium">🎯</span> Match my resume to a job
-                      </button>
-                      <button
-                        onClick={() => setInputText("What skills should I highlight?")}
-                        className="group rounded-lg border border-slate-200 bg-slate-50 px-4 py-3 text-left text-sm text-slate-700 transition-all hover:border-indigo-300 hover:bg-indigo-50 hover:text-indigo-700 dark:border-slate-700 dark:bg-slate-800/50 dark:text-slate-300 dark:hover:border-indigo-600 dark:hover:bg-indigo-950/30 dark:hover:text-indigo-400"
-                      >
-                        <span className="font-medium">⭐</span> What skills should I highlight?
-                      </button>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            ) : (
-              /* Chat Layout - When messages exist */
-              <div className="flex flex-1 flex-col rounded-xl border border-slate-200 bg-white shadow-sm dark:border-slate-700 dark:bg-slate-800">
-                {/* Chat Messages Area */}
-                <div className="flex-1 overflow-y-auto p-6">
-                  <div className="mx-auto max-w-3xl">
-                    {/* Chat Messages */}
-                    <div className="space-y-4">
-                      {messages.map((message) => (
-                        <div
-                          key={message.id}
-                          className={`flex gap-4 ${
-                            message.role === 'user' ? 'justify-end' : 'justify-start'
-                          }`}
-                        >
-                          {message.role === 'assistant' && (
-                            <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-gradient-to-br from-indigo-600 to-blue-600 dark:from-indigo-500 dark:to-blue-500">
-                              <svg className="h-5 w-5 text-white" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 10V3L4 14h7v7l9-11h-7z" />
-                              </svg>
-                            </div>
-                          )}
-                          <div
-                            className={`max-w-[80%] rounded-2xl px-4 py-3 ${
-                              message.role === 'user'
-                                ? 'bg-indigo-600 text-white dark:bg-indigo-500'
-                                : 'bg-slate-100 text-slate-900 dark:bg-slate-700 dark:text-slate-100'
-                            }`}
-                          >
-                            <p className="whitespace-pre-wrap text-sm leading-relaxed">{message.content}</p>
-                          </div>
-                          {message.role === 'user' && (
-                            <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-slate-200 dark:bg-slate-700">
-                              <svg className="h-5 w-5 text-slate-600 dark:text-slate-300" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
-                              </svg>
-                            </div>
-                          )}
-                        </div>
-                      ))}
-                      
-                      {/* Loading indicator */}
-                      {isLoading && (
-                        <div className="flex gap-4 justify-start">
-                          <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-gradient-to-br from-indigo-600 to-blue-600 dark:from-indigo-500 dark:to-blue-500">
-                            <svg className="h-5 w-5 text-white" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 10V3L4 14h7v7l9-11h-7z" />
-                            </svg>
-                          </div>
-                          <div className="rounded-2xl bg-slate-100 px-4 py-3 dark:bg-slate-700">
-                            <div className="flex gap-1">
-                              <div className="h-2 w-2 animate-bounce rounded-full bg-slate-400 [animation-delay:-0.3s]"></div>
-                              <div className="h-2 w-2 animate-bounce rounded-full bg-slate-400 [animation-delay:-0.15s]"></div>
-                              <div className="h-2 w-2 animate-bounce rounded-full bg-slate-400"></div>
-                            </div>
-                          </div>
-                        </div>
-                      )}
-                      <div ref={messagesEndRef} />
-                    </div>
-                  </div>
-                </div>
-
-                {/* Input Area */}
-                <div className="border-t border-slate-200 bg-slate-50/50 p-4 dark:border-slate-700 dark:bg-slate-900/50">
-                  <div className="mx-auto max-w-3xl">
-                    <div className="flex items-end gap-3">
-                      {/* Text Input */}
-                      <div className="flex-1">
-                        <div className="rounded-lg border border-slate-300 bg-white p-3 shadow-sm dark:border-slate-600 dark:bg-slate-700">
-                          <div className="mb-2 flex items-center gap-2">
-                            <svg className="h-4 w-4 text-slate-400 dark:text-slate-500" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13.828 10.172a4 4 0 00-5.656 0l-4 4a4 4 0 105.656 5.656l1.102-1.101m-.758-4.899a4 4 0 005.656 0l4-4a4 4 0 00-5.656-5.656l-1.1 1.1" />
-                            </svg>
-                            <span className="text-sm text-slate-500 dark:text-slate-400">
-                              Paste job description or URL here
-                            </span>
-                          </div>
-                          <textarea
-                            ref={textareaRef}
-                            value={inputText}
-                            onChange={(e) => setInputText(e.target.value)}
-                            onKeyDown={handleKeyDown}
-                            placeholder="Type your message... (Press Enter to send, Shift+Enter for new line)"
-                            className="w-full resize-none border-0 bg-transparent text-sm text-slate-900 placeholder-slate-400 focus:outline-none focus:ring-0 dark:text-slate-100 dark:placeholder-slate-500"
-                            rows={1}
-                            style={{ minHeight: '24px', maxHeight: '200px' }}
-                          />
-                        </div>
-                      </div>
-                      
-                      {/* Send Button */}
-                      <button
-                        onClick={handleSendMessage}
-                        disabled={!inputText.trim() || isLoading}
-                        className={`flex h-12 w-12 items-center justify-center rounded-lg bg-gradient-to-br from-indigo-600 to-blue-600 text-white shadow-md transition-all hover:shadow-lg disabled:cursor-not-allowed disabled:opacity-50 dark:from-indigo-500 dark:to-blue-500 ${
-                          inputText.trim() && !isLoading ? 'hover:scale-105' : ''
-                        }`}
-                        aria-label="Send message"
-                      >
-                        {isLoading ? (
-                          <svg className="h-5 w-5 animate-spin" fill="none" viewBox="0 0 24 24">
-                            <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                            <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                          </svg>
-                        ) : (
-                          <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 19l9 2-9-18-9 18 9-2zm0 0v-8" />
-                          </svg>
-                        )}
-                      </button>
-                    </div>
-                    <p className="mt-2 text-xs text-slate-500 dark:text-slate-400">
-                      Powered by AI. Your data is processed securely and privately.
-                    </p>
-                  </div>
-                </div>
-              </div>
-            )}
+              onSendMessage={handleSendMessage}
+              onClickUpload={handleClickUpload}
+            />
           </div>
         </div>
       </main>
