+++ frontend/src/components/Chat/ChatContainer.tsx
@@ -1,0 +1,87 @@
+import { Message } from '@/types';
+import { ChatMessage } from './ChatMessage';
+import { ChatInput } from './ChatInput';
+import { EmptyState } from './EmptyState';
+
+interface ChatContainerProps {
+  messages: Message[];
+  inputText: string;
+  isLoading: boolean;
+  textareaRef: React.RefObject<HTMLTextAreaElement | null>;
+  messagesEndRef: React.RefObject<HTMLDivElement | null>;
+  onInputChange: (value: string) => void;
+  onKeyDown: (e: React.KeyboardEvent<HTMLTextAreaElement>) => void;
+  onSendMessage: () => void;
+  onClickUpload: () => void;
+}
+
+export const ChatContainer = ({
+  messages,
+  inputText,
+  isLoading,
+  textareaRef,
+  messagesEndRef,
+  onInputChange,
+  onKeyDown,
+  onSendMessage,
+  onClickUpload,
+}: ChatContainerProps) => {
+  if (messages.length === 0) {
+    return (
+      <EmptyState
+        textareaRef={textareaRef}
+        inputText={inputText}
+        isLoading={isLoading}
+        onInputChange={onInputChange}
+        onKeyDown={onKeyDown}
+        onSendMessage={onSendMessage}
+        onClickUpload={onClickUpload}
+      />
+    );
+  }
+
+  return (
+    <div className="flex flex-1 flex-col rounded-xl border border-slate-200 bg-white shadow-sm dark:border-slate-700 dark:bg-slate-800">
+      {/* Chat Messages Area */}
+      <div className="flex-1 overflow-y-auto p-6">
+        <div className="mx-auto max-w-3xl">
+          <div className="space-y-4">
+            {messages.map((message) => (
+              <ChatMessage key={message.id} message={message} />
+            ))}
+            
+            {/* Loading indicator */}
+            {isLoading && (
+              <div className="flex gap-4 justify-start">
+                <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-gradient-to-br from-indigo-600 to-blue-600 dark:from-indigo-500 dark:to-blue-500">
+                  <svg className="h-5 w-5 text-white" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 10V3L4 14h7v7l9-11h-7z" />
+                  </svg>
+                </div>
+                <div className="rounded-2xl bg-slate-100 px-4 py-3 dark:bg-slate-700">
+                  <div className="flex gap-1">
+                    <div className="h-2 w-2 animate-bounce rounded-full bg-slate-400 [animation-delay:-0.3s]"></div>
+                    <div className="h-2 w-2 animate-bounce rounded-full bg-slate-400 [animation-delay:-0.15s]"></div>
+                    <div className="h-2 w-2 animate-bounce rounded-full bg-slate-400"></div>
+                  </div>
+                </div>
+              </div>
+            )}
+            <div ref={messagesEndRef} />
+          </div>
+        </div>
+      </div>
+
+      {/* Input Area */}
+      <ChatInput
+        textareaRef={textareaRef}
+        inputText={inputText}
+        isLoading={isLoading}
+        onInputChange={onInputChange}
+        onKeyDown={onKeyDown}
+        onSendMessage={onSendMessage}
+      />
+    </div>
+  );
+};
+
