+++ frontend/src/components/Chat/ChatMessage.tsx
@@ -1,0 +1,41 @@
+import { Message } from '@/types';
+
+interface ChatMessageProps {
+  message: Message;
+}
+
+export const ChatMessage = ({ message }: ChatMessageProps) => {
+  return (
+    <div
+      className={`flex gap-4 ${
+        message.role === 'user' ? 'justify-end' : 'justify-start'
+      }`}
+    >
+      {message.role === 'assistant' && (
+        <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-gradient-to-br from-indigo-600 to-blue-600 dark:from-indigo-500 dark:to-blue-500">
+          <svg className="h-5 w-5 text-white" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 10V3L4 14h7v7l9-11h-7z" />
+          </svg>
+        </div>
+      )}
+      <div
+        className={`max-w-[80%] rounded-2xl px-4 py-3 ${
+          message.role === 'user'
+            ? 'bg-indigo-600 text-white dark:bg-indigo-500'
+            : 'bg-slate-100 text-slate-900 dark:bg-slate-700 dark:text-slate-100'
+        }`}
+      >
+        <p className="whitespace-pre-wrap text-sm leading-relaxed">{message.content}</p>
+      </div>
+      {message.role === 'user' && (
+        <div className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-full bg-slate-200 dark:bg-slate-700">
+          <svg className="h-5 w-5 text-slate-600 dark:text-slate-300" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
+          </svg>
+        </div>
+      )}
+    </div>
+  );
+};
+
+
