+++ frontend/src/components/ResumeUpload.tsx
@@ -1,0 +1,156 @@
+import { formatFileSize } from '@/lib/utils';
+
+interface ResumeUploadProps {
+  uploadedFile: File | null;
+  isDragging: boolean;
+  uploadError: string | null;
+  fileInputRef: React.RefObject<HTMLInputElement | null>;
+  onDragOver: (e: React.DragEvent) => void;
+  onDragLeave: (e: React.DragEvent) => void;
+  onDrop: (e: React.DragEvent) => void;
+  onFileInputChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onClickUpload: () => void;
+  onRemoveFile: () => void;
+}
+
+export const ResumeUpload = ({
+  uploadedFile,
+  isDragging,
+  uploadError,
+  fileInputRef,
+  onDragOver,
+  onDragLeave,
+  onDrop,
+  onFileInputChange,
+  onClickUpload,
+  onRemoveFile,
+}: ResumeUploadProps) => {
+  return (
+    <div className="rounded-xl border border-slate-200 bg-white p-6 shadow-sm dark:border-slate-700 dark:bg-slate-800">
+      <h2 className="mb-4 text-lg font-semibold text-slate-900 dark:text-slate-100">
+        Upload Resume
+      </h2>
+      
+      {/* Hidden file input */}
+      <input
+        ref={fileInputRef}
+        type="file"
+        accept=".pdf,application/pdf"
+        onChange={onFileInputChange}
+        className="hidden"
+      />
+      
+      {/* PDF Upload Area */}
+      {!uploadedFile ? (
+        <div
+          onDragOver={onDragOver}
+          onDragLeave={onDragLeave}
+          onDrop={onDrop}
+          onClick={onClickUpload}
+          className={`flex h-40 cursor-pointer flex-col items-center justify-center rounded-lg border-2 border-dashed p-6 text-center transition-all ${
+            isDragging
+              ? 'border-indigo-500 bg-indigo-100 dark:border-indigo-400 dark:bg-indigo-950/50'
+              : 'border-slate-300 bg-slate-50 hover:border-indigo-400 hover:bg-indigo-50/50 dark:border-slate-600 dark:bg-slate-700/50 dark:hover:border-indigo-500 dark:hover:bg-indigo-950/30'
+          }`}
+        >
+          <svg
+            className={`mb-3 h-10 w-10 transition-colors ${
+              isDragging
+                ? 'text-indigo-600 dark:text-indigo-400'
+                : 'text-slate-400 dark:text-slate-500'
+            }`}
+            fill="none"
+            viewBox="0 0 24 24"
+            stroke="currentColor"
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth={2}
+              d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12"
+            />
+          </svg>
+          <p className="text-sm font-medium text-slate-600 dark:text-slate-300">
+            {isDragging ? 'Drop PDF here' : 'Drop PDF here or click to browse'}
+          </p>
+          <p className="mt-1 text-xs text-slate-500 dark:text-slate-400">
+            Max file size: 10MB
+          </p>
+        </div>
+      ) : (
+        <div className="rounded-lg border border-slate-200 bg-slate-50 p-4 dark:border-slate-700 dark:bg-slate-700/50">
+          <div className="flex items-start gap-3">
+            <div className="flex h-10 w-10 flex-shrink-0 items-center justify-center rounded-lg bg-indigo-100 dark:bg-indigo-900/50">
+              <svg
+                className="h-6 w-6 text-indigo-600 dark:text-indigo-400"
+                fill="none"
+                viewBox="0 0 24 24"
+                stroke="currentColor"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth={2}
+                  d="M7 21h10a2 2 0 002-2V9.414a1 1 0 00-.293-.707l-5.414-5.414A1 1 0 0012.586 3H7a2 2 0 00-2 2v14a2 2 0 002 2z"
+                />
+              </svg>
+            </div>
+            <div className="min-w-0 flex-1">
+              <p className="truncate text-sm font-medium text-slate-900 dark:text-slate-100">
+                {uploadedFile.name}
+              </p>
+              <p className="mt-1 text-xs text-slate-500 dark:text-slate-400">
+                {formatFileSize(uploadedFile.size)}
+              </p>
+            </div>
+            <button
+              onClick={(e) => {
+                e.stopPropagation();
+                onRemoveFile();
+              }}
+              className="flex-shrink-0 rounded-lg p-1 text-slate-400 transition-colors hover:bg-slate-200 hover:text-slate-600 dark:hover:bg-slate-600 dark:hover:text-slate-300"
+              aria-label="Remove file"
+            >
+              <svg
+                className="h-5 w-5"
+                fill="none"
+                viewBox="0 0 24 24"
+                stroke="currentColor"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth={2}
+                  d="M6 18L18 6M6 6l12 12"
+                />
+              </svg>
+            </button>
+          </div>
+        </div>
+      )}
+      
+      {/* Error Message */}
+      {uploadError && (
+        <div className="mt-3 rounded-lg border border-red-200 bg-red-50 p-3 dark:border-red-900/50 dark:bg-red-950/30">
+          <div className="flex items-start gap-2">
+            <svg
+              className="h-5 w-5 flex-shrink-0 text-red-600 dark:text-red-400"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth={2}
+                d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
+              />
+            </svg>
+            <p className="text-sm text-red-700 dark:text-red-400">{uploadError}</p>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
+
