+++ frontend/src/types/index.ts
@@ -1,0 +1,8 @@
+export interface Message {
+  id: string;
+  role: 'user' | 'assistant';
+  content: string;
+  timestamp: Date;
+}
+
+
